trade_management/requirements.txt
django >=3.0
djangorestframework
marshmallow
pika
bintrees